Post Merge
RAM:   [==        ]  18.1% (used 59376 bytes from 327680 bytes)
Flash: [========= ]  87.4% (used 1661121 bytes from 1900544 bytes)

initSelect int->uint8_t

+++ src/App/LedEffects.h
@@ -2924,7 +2924,7 @@ class LiveScriptEffect: public Effect {
   
   void controls(LedsLayer &leds, JsonObject parentVar) {
     Effect::controls(leds, parentVar);
-    ui->initSelect(parentVar, "script", UINT16_MAX, false , [&leds](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
+    ui->initSelect(parentVar, "script", UINT8_MAX, false , [&leds](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
       case onUI: {
         // ui->setComment(var, "Fixture to display effect on");
         JsonArray options = ui->setOptions(var);
