+++ src/User/UserModLive.h
@@ -145,7 +145,7 @@ public:
 
     parentVar = ui->initUserMod(parentVar, name, 6310);
 
-    ui->initSelect(parentVar, "script", UINT16_MAX, false, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
+    ui->initSelect(parentVar, "script", UINT8_MAX, false, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
       case onUI: {
         // ui->setComment(var, "Fixture to display effect on");
         JsonArray options = ui->setOptions(var);
